% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gCrisprTools.R
\docType{data}
\name{aln}
\alias{aln}
\title{Precalculated alignment statistics of a crispr screen}
\source{
Genentech, Inc.
}
\description{
Example alignment matrix file for the provided example Crispr screen.
All sample, gRNA, and Gene information has been anonymized and randomized.
}
\examples{
data('aln')
head(aln)
}
\seealso{
Please see \file{vignettes/Crispr_example_workflow.R} for details.
}
