% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gCrisprTools.R
\docType{data}
\name{ann}
\alias{ann}
\title{Annotation file for a mouse Crispr library}
\source{
Genentech, Inc.
}
\description{
Example annotation file for the screen data provided in \code{es}. 
All sample, gRNA, and Gene information has been anonymized and randomized.
}
\examples{
data('ann')
head(ann)
}
\seealso{
Please see \file{vignettes/Crispr_example_workflow.R} for details.
}
