% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReports.R
\name{appendDateAndExt}
\alias{appendDateAndExt}
\title{Add formatted timestamp and extension to a file name}
\usage{
appendDateAndExt(name, ext)
}
\arguments{
\item{name}{character}

\item{ext}{character - extension including a ".".}
}
\value{
character
}
\description{
Add formatted timestamp and extension to a file name
}
\keyword{internal}
