% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareContrasts.R
\name{ct.compareContrasts}
\alias{ct.compareContrasts}
\title{Identify Replicated Signals in Pooled Screens Using Conditional Scoring}
\usage{
ct.compareContrasts(
  mainresult,
  validationresult,
  statistics = c("best.q", "best.p"),
  cutoffs = c(0.1, 0.1),
  same.dir = TRUE,
  return.stats = FALSE,
  ...
)
}
\arguments{
\item{mainresult}{A (possibly simplified) results data.frame containing the main contrast to be analyzed. See \code{\link{ct.generateResults}}.}

\item{validationresult}{Column of the provided \code{summaryDF} to consider. Must be \code{geneID} or \code{geneSymbol}.}

\item{statistics}{Statistics to use to define congruence; may be a single value, but internally coerced to a vector of length 2 where the first 
value corresponds to the stringent cutoff annd the second value is used for the relaxed cutoff. Must be "best.p" or "best.q".}

\item{cutoffs}{Numeric value(s) corresponding to the significance cutoff(s) used to define stringent and relaxed values of `statistics`. 
Internally coerced to a vector of length 2.}

\item{same.dir}{Logical indicating whether replicating signals are expected to go in the same direction (e.g., enrich/deplete in both screens)}

\item{return.stats}{When TRUE, return the significance of overlap instead of the logical vector (via a Hypergeometric Test).}

\item{...}{Other arguments to \link{`ct.simpleResult`}, especially `collapse`.}
}
\value{
If `return.stats` is `FALSE`, returns the simplified `mainresults` data.frame, with a `replicated` logical column indicating whether a 
signal replicates. Incomparable elements (e.g., targets not assayed in the provided `validationresult`) are set to `NA`.  If `return.stats` is 
`TRUE`, returns a named list indicating the hypergeometric test *P*-values summarizing the evidence for significantly enriched signal 
replication across screens,
}
\description{
This function identifies signals in a provided screening experiment contrast that are also present in a 
second contrast sing a conditional strategy. Specifically, this function identifies all significant signals (according to user definitions) 
in a provided results DF and returns a `simplifiedResult` dataframe with an appended logical column indicating whether there 
is evidence for signal replication in a second provided resultsDF. 

Signals are considered replicated in the second screen contrast if the best signal associated with the same same target is similarly enriched 
or depleted in the replicating screen, and the target signal is associated with a P-value at or below a user-indicated threshold. 

Signals are compared across screens on the basis of \code{\link{ct.regularizeContrasts}}, so users must provide an identifier 
with which to standardize targets across the specified contrasts (`geneID` by default).
}
\examples{
data('resultsDF')
summary(ct.compareContrasts(resultsDF, resultsDF[1:5000,])$replicated)
ct.compareContrasts(resultsDF, resultsDF[1:5000,], return.stats = TRUE)
}
\author{
Russell Bainer
}
