% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewGuides.R
\name{ct.drawColorLegend}
\alias{ct.drawColorLegend}
\title{Draw a density color legend.}
\usage{
ct.drawColorLegend(dens, colorscale)
}
\arguments{
\item{dens}{A density object.}

\item{colorscale}{A vector of colors to draw behind the density.}
}
\value{
A color legend on the current graphics device.
}
\description{
This is a function called internally by ct.viewGuides to generate the color legend. End users should not use it.
}
\author{
Russell Bainer
}
\keyword{internal}
