% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuideCDF.R
\name{ct.ecdf}
\alias{ct.ecdf}
\title{Generate a cumulative tally of reads by guide rank}
\usage{
ct.ecdf(vector)
}
\arguments{
\item{vector}{An input numeric vector to be aggregated.}
}
\value{
A numeric vector of the cumulative sum
}
\description{
This function returns a numeric vector of the same length as the input, where each element \code{n} contains the proportion
the sum of the full vector that is captured by its first \code{1-n} elements (arranged in descending order).
}
\author{
Russell Bainer
}
\keyword{internal}
