% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileutilities.R
\name{ct.extractSE}
\alias{ct.extractSE}
\title{Extract gCrisprTools objects from a `SummarizedExperiment`}
\usage{
ct.extractSE(what, se)
}
\arguments{
\item{what}{What gCrisprTools-friendly object to compile. options are: 
- `es`: an `ExpressionSet`, compiled from the `exprs` slot of the `assayData`, with `colData` and `rowData` saved as the `fData` and `pData`
- `ann`: a gCrisprTools annotation from the `rowData`}

\item{se}{The `SummarizedExperiment` object.}
}
\value{
The specified gCrisprTools-friendly object
}
\description{
Utility function to enable gCrisprTools functions to take `SummarizedExperiment` class objects (or subclasses thereof) as input. 
Throws errors when unexpected things happen.
}
\author{
Russell Bainer
}
