% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRAalpha.R
\name{ct.makeRhoNull}
\alias{ct.makeRhoNull}
\title{Make null distribution for RRAalpha tests}
\usage{
ct.makeRhoNull(n, p, nperm)
}
\arguments{
\item{n}{single integer, number of guides per gene}

\item{p}{numeric vector of rank statistics}

\item{nperm}{single integer, how many random samples to take.}
}
\value{
numeric vector of Rho values
}
\description{
Makes random distribution of Rho value by taking nperm random samples of n rank stats, p.
}
\examples{
ct.makeRhoNull(3, 1:9, 5)
}
