% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREATdb.R
\name{ct.mgseaPrep}
\alias{ct.mgseaPrep}
\title{Prepare a resultsDF object for analysis via MultiGSEA}
\usage{
ct.mgseaPrep(
  dflist,
  cutoff = 0.1,
  statistic = c("best.q", "best.p"),
  gdb = NULL,
  ...
)
}
\arguments{
\item{dflist}{A list of gCrisprTools results `data.frames` to be formatted.}

\item{cutoff}{Numeric maximum value of `statistic` to define significance.}

\item{statistic}{Should cutoffs be calculated based on FDR (`best.q`) or P-value (`best.p`)?}

\item{gdb}{Optionally, a `GeneSetDb` object to enable proper registration of the output. If provided, the 
collapsing features in the provided `simpleDF`s must be present in the `gsd@db$feature_id` slot.}

\item{...}{Other parameters to lower functions, especially `ct.simpleResult()`.}
}
\value{
A list of `data.frames` formatted for evaluation with `multiGSEA`.
}
\description{
Prepare a resultsDF object for analysis via MultiGSEA
}
