% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREATdb.R
\name{ct.multiGSEA}
\alias{ct.multiGSEA}
\title{Geneset Enrichment within a CRISPR screen using multiGSEA

This function identifies differentially enriched/depleted ontological categories within the hits of a CRISPR screen 
given a provided `GenseSetDb()` and a results `data.frame` created by `ct.generateResults()`. Testing is performed using 
a Hypergeometric test, and results are returned as a `MultiGSEAResult` object defined in the `multiGSEA` package. Note that the 
`@logFC` slot in the returned object will contain the median gRNA lfc across all associated guides, which in some cases may 
have dubious interpretive value. 

This method used overrepresentation analysis, derived from `limma::kegga()`, and incorporates the number of gRNAs associated with 
each Target (inferred from the `geneSymbol` column of the `resultsDF`)  as the bias vector (because standard aggregation methods 
should be underpowered for targets with few guides). Setting `unbiased` = `TRUE` suppresses this behavior, which is identical to 
a hypergeometric test.}
\usage{
ct.multiGSEA(
  resultsDF,
  gsdb,
  cutoff = 0.1,
  stat = c("q", "p", "rho"),
  unbiased = FALSE
)
}
\arguments{
\item{resultsDF}{`data.frame` returned by `ct.generateResults()`.}

\item{gsdb}{`GenseSetDb` object containing annotations.}

\item{cutoff}{Q, P, or Rho statistic cutoff defining significant enrichment/depletion in the screen. Default is 0.1.}

\item{stat}{Statistic to be used in calling enrichment/depletion in the screen. Must be one of 'q', 'p', or 'rho'.}

\item{unbiased}{Logical indicating whether to estimate bias on the basis of the number of gRNAs associated with each target.}
}
\description{
Geneset Enrichment within a CRISPR screen using multiGSEA

This function identifies differentially enriched/depleted ontological categories within the hits of a CRISPR screen 
given a provided `GenseSetDb()` and a results `data.frame` created by `ct.generateResults()`. Testing is performed using 
a Hypergeometric test, and results are returned as a `MultiGSEAResult` object defined in the `multiGSEA` package. Note that the 
`@logFC` slot in the returned object will contain the median gRNA lfc across all associated guides, which in some cases may 
have dubious interpretive value. 

This method used overrepresentation analysis, derived from `limma::kegga()`, and incorporates the number of gRNAs associated with 
each Target (inferred from the `geneSymbol` column of the `resultsDF`)  as the bias vector (because standard aggregation methods 
should be underpowered for targets with few guides). Setting `unbiased` = `TRUE` suppresses this behavior, which is identical to 
a hypergeometric test.
}
\examples{
data('ann')
data('resultsDF')
#gsd <- multiGSEA::getMSigGeneSetDb(c('h', 'c2'), 'mouse', id.type = 'entrez')

}
\author{
Steve Lianoglou for multiGSEA; Russell Bainer for wrapping functions.
}
