% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRAalpha.R
\name{ct.numcores}
\alias{ct.numcores}
\title{Checks and Possibly Sets the Number of Cores to be Used in Parallel Processing}
\usage{
ct.numcores()
}
\value{
Nothing, but invisibly sets \code{options(mc.cores)} if currently \code{NULL}.
}
\description{
This function determines the number of cores that the user is expecting to 
use during parallel processing operations, and if absent, sets the \code{mc.cores} option
to the maximum value. Users who do not wish to use all available cores during parallel 
processing should do so by invoking \code{options()} from the command line prior to analysis.
}
\author{
Russell Bainer, Pete Haverty
}
\keyword{internal}
