% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crispr_Enrichment_test.R
\name{ct.PantherPathwayEnrichment}
\alias{ct.PantherPathwayEnrichment}
\title{Run a (limited) Pathway Enrichment Analysis on the results of a Crispr experiment.}
\usage{
ct.PantherPathwayEnrichment(
  summaryDF,
  pvalue.cutoff = 0.01,
  enrich = TRUE,
  organism = "human",
  db.cut = 10
)
}
\arguments{
\item{summaryDF}{A dataframe summarizing the results of the screen, returned by the function \code{\link{ct.generateResults}}.}

\item{pvalue.cutoff}{A gene-level p-value cutoff defining targets of interest within the screen. Note that this is a
nominal p-value cutoff to preserve end-user flexibility.}

\item{enrich}{Logical indicating whether to consider guides that are enriched (default) or depleted within the screen.}

\item{organism}{The species of the cell line used in the screen; currently only 'human' or 'mouse' are supported.}

\item{db.cut}{Minimum number of genes annotated to a given to a pathway within the screen in order to consider it in the enrichment test.}
}
\value{
A dataframe of enriched pathways.
}
\description{
This function enables some limited geneset enrichment-type analysis of data derived from a pooled Crispr
screen using the PANTHER pathway database. Specifically, it identifies the set of targets significantly enriched or
depleted in a \code{summaryDF} object returned from \code{ct.generateResults} and compares that set to the remaining
targets in the screening library using a hypergeometric test.

Note that many Crispr gRNA libraries specifically target biased sets of genes, often focusing on genes involved
in a particular pathway or encoding proteins with a shared biological property. Consequently, the enrichment results
returned by this function represent the pathways containing genes disproportionately targeted *within the context
of the screen*, and may or may not be informative of the underlying biology in question. This means that
pathways not targeted by a Crispr library will obviously never be enriched within the positive target set regardless of
their biological relevance, and pathways enriched within a focused library screen are similarly expected to partially
reflect the composition of the library and other confounding issues (e.g., number of targets within a pathway).
Analysts should therefore use this function with care. For example, it might be unsurprising to detect pathways related
to histone modification within a screen employing a crispr library targeting epigenetic regulators.
}
\author{
Russell Bainer, Steve Lianoglou
}
\keyword{internal}
