% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRAalpha.R
\name{ct.RRAalpha}
\alias{ct.RRAalpha}
\title{Aggregation of P-value Ranks using a Beta Distribution and Alpha Cutoff}
\usage{
ct.RRAalpha(p, g.key, alpha, shuffle = FALSE, return.obj = TRUE)
}
\arguments{
\item{p}{A single column matrix of rank statistics, with \code{row.names} indicating the gRNA labels.}

\item{g.key}{data.frame with guide and gene names}

\item{alpha}{The alpha cutoff parameter, corresponding to the P-value threshold or fold change proportion at which gRNAs should no longer be considered to be 
differentially expressed. Alternatively, this can be provided as a logical vector of the same length as the number of rows in \code{p}, 
containing only \code{TRUE} and \code{FALSE} elements indicating whether the element should be included during the aggeregation step.}

\item{shuffle}{Logical indicating whether to shuffle the rank statistics prior to calculating the rho statistics (useful for permutation).}

\item{return.obj}{Name of the environment to record results in, or \code{TRUE} to return the RRAa results directly as a numeric vector of genewise rho statistics. 
If an environment is supplied, the function will directly increment the \code{target.positive.iterations} variable within it on the basis of the \code{obs} 
variable in the specified \code{environment}.}
}
\value{
Nothing, or a named list of target-level P-values, which are treated as a rho statistic in the permutation step.
}
\description{
This function is called internally as a single instance of the beta aggregation step in RRAa. Users should not interact with it directly. 
The expected input is a list of rank statistics, and a paired \code{alpha} argument defining which values to consider in downstream analyses (see below).
}
\examples{
data('fit')
data('ann')
geneScores <- ct.RRAalpha(fit$p.value, ann, alpha = 0.1, shuffle = FALSE, return.obj = TRUE)
}
\author{
Russell Bainer
}

