% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRAalpha.R
\name{ct.RRAaPvals}
\alias{ct.RRAaPvals}
\title{gRNA signal aggregation via RRAa, optionally using multiple cores.}
\usage{
ct.RRAaPvals(p, g.key, alpha, permute, multicore = TRUE, core.perm = 100,
  permutation.seed = NULL)
}
\arguments{
\item{p}{A single column matrix of ranking scores, with row.names indicating the gRNA labels}

\item{g.key}{An annotation data frame of gRNAs, minimally containing a factorized "geneSymbol" column indicating the target names. This is typically generated by calling the \code{ct.buildKeyFromAnnotation()} function.}

\item{alpha}{The alpha cutoff parameter, corresponding to the P-value threshold or fold change proportion at which gRNAs should no longer be considered to be 
differentially expressed. Alternatively, this can be provided as a logical vector of the same length as the number of rows in \code{p}, 
containing only \code{TRUE} and \code{FALSE} elements indicating whether the element should be included during the aggeregation step.}

\item{permute}{Number of permutations to be used during empirical p-value estimation. In a multicore context the exact number of permutations may vary somewhat
to accomodate the corresponding system archetecture but should be close to the specified permutation number.}

\item{multicore}{Logical indicating whether to use multiple cores to calculate p-values.}

\item{core.perm}{Maximum number of permutations to run on each core (only relevant when \code{multicore} is \code{TRUE}).}

\item{permutation.seed}{numeric seed for permutation reproducibility.
Default: \code{NULL} means to not set any seed.}
}
\value{
A named list of target-level empirical P-values.
}
\description{
This is a wrapper function implementing the RRAalpha p-value aggregation algorithm. Takes in a set of gRNA rank scores (formatted as a single-column 
numeric matrix with row.names indicating the guide names) and a list object of gRNA annotations (names are the gene targets, and each element of the list contains 
a vector of the corresponding guide names). The rank scores are converted to gene-level statistics that are thenm transformed into empirical p-values by permutation.
}
\examples{
data('fit')
data('ann')
genePvals <- ct.RRAaPvals(fit$p.value, ann, alpha = 0.1, permute = 100, multicore = FALSE)
genePvals <- ct.RRAaPvals(fit$p.value, ann, alpha = 0.1, permute = 100, multicore = TRUE, core.perm = 10)
}
\author{
Russell Bainer
}
\keyword{internal}

