% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileutilities.R
\name{ct.softLog}
\alias{ct.softLog}
\title{Log10 transform empirical P-values with a pseudocount}
\usage{
ct.softLog(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
-log10-transformed version of X.
}
\description{
This function -log10 transforms empirical P-values by adding a pseudocount of 1/2 the minimum nonzero value.
}
