% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gCrisprTools.R
\docType{data}
\name{essential.genes}
\alias{essential.genes}
\title{Artificial list of 'essential' genes in the example Crispr screen 
included for plotting purposes}
\source{
Russell Bainer
}
\description{
Example gene list, designed to demonstrate ROC and PRC functions
All sample, gRNA, and Gene information has been anonymized and randomized.
}
\examples{
data("essential.genes")
essential.genes
}
\seealso{
Please see \file{vignettes/Crispr_example_workflow.R} for details.
}
