% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gCrisprTools.R
\docType{data}
\name{fit}
\alias{fit}
\title{Precalculated contrast fit from a Crispr screen}
\source{
Genentech, Inc.
}
\description{
A precalculated fit object (class \code{MArrayLM}) comparing the death 
and control expansion arms of a crispr screen performed at Genentech, Inc. 
All sample, gRNA, and Gene information has been anonymized and randomized.
}
\examples{
data('fit')
show(fit)
}
\seealso{
Please see \file{vignettes/Crispr_example_workflow.R} for model details.
}
