% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileutilities.R
\name{initOutDir}
\alias{initOutDir}
\title{Initializes the output directory}
\usage{
initOutDir(outdir)
}
\arguments{
\item{outdir}{character vector pointing to a directory to check/create}
}
\value{
TRUE if the output directory was created, otherwise FALSE (it might
already exist).
}
\description{
If outdir is NULL, then no directory is checked/created. This also implies
that creating plots is not possible.
}
\author{
Steve Lianoglou, Russell Bainer
}

