% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReports.R
\name{renderReport}
\alias{renderReport}
\title{Internal wrapper to generate html markdown reports from existing templates}
\usage{
renderReport(reportNameBase, templateName, rmdParamList, outdir = NULL)
}
\arguments{
\item{reportNameBase}{character - name of the report's file.}

\item{templateName}{character - name of the rmarkdown template file in the
standard location. Passed through to \code{template} argument of the
\code{\link[rmarkdown]{draft}}.}

\item{rmdParamList}{list of named report parameters. Passed through to
\code{params} argument of the \code{\link[rmarkdown]{render}}.}

\item{outdir}{An optional character string indicating the directory in which to 
generate the report. If \code{NULL}, a temporary directory will be automatically generated.}
}
\value{
character with a path to html report in the temporary directory.
}
\description{
Internal wrapper to generate html markdown reports from existing templates
}
\keyword{internal}
