% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gCrisprTools.R
\docType{data}
\name{resultsDF}
\alias{resultsDF}
\title{Precalculated gene-level summary of a crispr screen}
\source{
Genentech, Inc.
}
\description{
A precalculated summary Dataframe comparing the death and control expansion arms of 
the provided example Crispr screen (using 8 cores, seed = 2). 
All sample, gRNA, and Gene information has been anonymized and randomized.
}
\examples{
data('resultsDF')
head(resultsDF)
}
\seealso{
Please see \file{vignettes/Crispr_example_workflow.R} for model details.
}
