% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gCrisprTools.R
\docType{data}
\name{se}
\alias{se}
\title{SummarizedExperiment of count data from a Crispr screen with strong selection}
\source{
Genentech, Inc.
}
\description{
SummarizedExperiment of raw counts from a screen in mouse cells performed at Genentech, Inc. 
All sample, gRNA, and Gene information has been anonymized and randomized.
}
\examples{
data('se')
print(se)
}
\seealso{
Please see \file{vignettes/Crispr_example_workflow.R} for details.
}
