% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREATdb.R
\name{ct.GREATdb}
\alias{ct.GREATdb}
\title{Update a gene-centric msdb object for GREAT-style enrichment analysis using a specified CRISPR annotation.}
\usage{
ct.GREATdb(
  annotation,
  gsdb = getMSigGeneSetDb(c("h", "c2"), "human"),
  minsize = 10
)
}
\arguments{
\item{annotation}{an annotation object returned by \code{ct.prepareAnnotation()}.}

\item{gsdb}{A gene-centric \code{GeneSetDb} object to conform to the relevant peakwise dataset.}

\item{minsize}{Minimum number of targets required to consider a geneset valid for analysis.}
}
\value{
A new \code{GeneSetDb} object with the features annotated genewise to pathways.
}
\description{
Update a gene-centric msdb object for GREAT-style enrichment analysis using a specified CRISPR annotation.
}
