% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{ct.normalizeGenewise}
\alias{ct.normalizeGenewise}
\title{Normalize sample abundance estimates genewise by the median values of a subset of guides}
\usage{
ct.normalizeGenewise(eset, annotation, toNorm, lib.size = NULL)
}
\arguments{
\item{eset}{An ExpressionSet object containing, at minimum, count data accessible by \code{exprs}.}

\item{annotation}{An annotation dataframe indicating the gene-level assignments of gRNAs in the geneSymbol column.}

\item{toNorm}{a logical vector indicating which gRNAs should be included in the normalization step.}

\item{lib.size}{An optional vector of voom-appropriate library size adjustment factors, usually calculated with \code{\link[edgeR]{calcNormFactors}} 
and transformed to reflect the appropriate library size. These adjustment factors are interpreted as the total library sizes for each sample, 
and if absent will be extrapolated from the columnwise count sums of the \code{exprs} slot of the \code{eset}.}
}
\value{
A normalized \code{eset}.
}
\description{
This function normalizes subsets of Crispr gRNA abundance estimates by equalizing the median 
abundances of a specified set of gRNAs annotated to the subset within each sample. The normalized values are returned as normalized counts in 
the '\code{exprs}' slot of the input eset. This normalization strategy is best suited for tiling screening experiments where the overall impact 
of a gRNA is imagined to be a combination of the gene-level and guide-level effects, which we take to be additive and independent.
}
\examples{
data('es')
data('ann')

#Build the sample key and library sizes for visualization
library(Biobase)
sk <- ordered(relevel(as.factor(pData(es)$TREATMENT_NAME), 'ControlReference'))
names(sk) <- row.names(pData(es))
ls <- colSums(exprs(es))

es.norm <- ct.normalizeGenewise(es, ann, toNorm = ifelse(rnorm(nrow(ann)) > 0, TRUE, FALSE), lib.size = ls)

ct.gRNARankByReplicate(es, sk, lib.size = ls)
ct.gRNARankByReplicate(es.norm, sk, lib.size = ls)
}
\author{
Russell Bainer
}
