% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crispr_Enrichment_test.R
\name{ct.targetSetEnrichment}
\alias{ct.targetSetEnrichment}
\title{Test Whether a Specified Target Set is Enriched Within a Pooled Screen}
\usage{
ct.targetSetEnrichment(summaryDF, targets, enrich = c(TRUE, FALSE),
  ignore = NULL)
}
\arguments{
\item{summaryDF}{A dataframe summarizing the results of the screen, returned by the function \code{\link{ct.generateResults}}.}

\item{targets}{A character vector containing the names of the targets to be tested. Only targets contained in the \code{geneID} 
column of the provided \code{summaryDF} are considered.}

\item{enrich}{Logical indicating whether to consider guides that are enriched (default) or depleted within the screen.}

\item{ignore}{Optionally, a character vector containing elements of the \code{geneID} column of the provided \code{summaryDF}
that should be ignored in the analysis (e.g., unassignable or nonfunctional targets, such as nontargeting controls). By default, 
this function omits targets with \code{geneSymbol} 'NoTarget'.}
}
\value{
A named list containing the tested target set and tables detailing the hypergeometric test results using various P-value and 
Q-value thresholds.
}
\description{
This function takes in a \code{resultsDF} and a vector of \code{targets} (contained in the \code{geneID} column of 
\code{resultsDF}) and determines whether the specified targets are enriched within the set of all significantly altered targets.  
It does this by iteratively testing whether \code{targets} are more likely to be among the set of enriched or depleted targets 
at various significance thresholds using a hypergeometric test. Note that the returned Hypergeometric P-values are not corrected 
for multiple testing.   

Returns a list detailing the \code{targets} used in the tests, and tables indicating the results of the hypergeometric test
at various significance thresholds.
}
\examples{
data(resultsDF)
tar <-  sample(unique(resultsDF$geneID), 20)
res <- ct.targetSetEnrichment(resultsDF, tar)
}
\author{
Russell Bainer
}
